+++ pages/index.tsx
@@ -1,11 +1,11 @@
-import { Grid } from '@mantine/core';
-import { getPerformance } from "firebase/performance";
+import {Grid} from '@mantine/core';
+import {getPerformance} from "firebase/performance";
 import Head from 'next/head';
-import Image, { ImageLoader } from 'next/image';
-import Link from 'next/link';
-import Ratings from '../components/ratings/Ratings';
 import firebaseApp from '../lib/firebase';
 import styles from '../styles/Home.module.css';
+import {useAuth} from "../lib/firebase.auth";
+import MovieCard from "../components/MovieCard/MovieCard";
+import React from "react";
 
 if (process.browser) {
   getPerformance(firebaseApp);
@@ -13,12 +13,14 @@ if (process.browser) {
 
 export default function Home({ posts }) {
 
+  const { user } = useAuth();
+
   return (
     <div className={styles.container}>
       <Head>
         <title>Five Stars</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
+        <meta name="description" content="Generated by create next app"/>
+        <link rel="icon" href="/favicon.ico"/>
       </Head>
 
       <main className={styles.main}>
@@ -29,15 +31,7 @@ export default function Home({ posts }) {
         <Grid gutter="lg">
           {posts.results.map(movie => (
             <Grid.Col sm={6} lg={4} key={movie.id}>
-              <div className={styles.card}>
-                <Link href={'/movie/' + movie.id} passHref>
-                  <h2>{movie.title}</h2>
-                </Link>
-                <Image loader={tmdbLoader} src={movie.poster_path} alt={movie.title} width={100} height={148} />
-                <div>
-                  <Ratings movie={movie} />
-                </div>
-              </div>
+              <MovieCard movie={movie} user={user}/>
             </Grid.Col>
           ))}
         </Grid>
@@ -52,8 +46,4 @@ export async function getStaticProps({ preview = null }) {
   return {
     props: { posts, preview },
   }
-}
-
-const tmdbLoader: ImageLoader = ({ src, width, quality }) => {
-  return `https://image.tmdb.org/t/p/w500/${src}?w=${width}&q=${quality || 75}`
 }
